requirements.txt
python-dotenv
requests
groq
google-genai
cohere